+++ packages/editor/web/src/RichContentEditor/RichContentEditor.jsx
@@ -319,6 +319,11 @@ class RichContentEditor extends Component {
     return 'handled';
   };
 
+  handleEscCommand = (_, event) => {
+    this.blur();
+    event?.preventDefault();
+  };
+
   getCustomCommandHandlers = () => ({
     commands: [
       ...this.pluginKeyBindings.commands,
@@ -342,7 +347,7 @@ class RichContentEditor extends Component {
       ...this.pluginKeyBindings.commandHandlers,
       tab: this.handleTabCommand,
       shiftTab: this.handleTabCommand,
-      esc: this.blur,
+      esc: this.handleEscCommand,
     },
   });