+++ CHANGELOG.md
@@ -19,6 +19,8 @@
 ### :bug: Bug Fix
 - `editor`
   - [#1410](https://github.com/wix-incubator/rich-content/pull/1410) fix space deletes atomic block by disabling keyboard inputs on atomic blocks
+- `editor`
+  - [#1381](https://github.com/wix-incubator/rich-content/pull/1381) fix blur editor on esc click
 </details>
 <hr/>